display score/pp for score buttons in songbrowser, depending on sorting type
also changed the default behavior a little bit, added a different convar to restore old behavior (scores_always_display_pp) to control it

TODO: would be nice to be able to sort ascending/descending, with an arrow to indicate sorting direction

+++ src/App/Neomod/HUD.cpp
@@ -685,7 +685,7 @@ void HUD::drawFps() {
     const f32 dpiScale = Osu::getUIScale();
 
     const i32 margin = std::round(3.0f * dpiScale);
-    const i32 shadowOffset = std::round(1.0f * dpiScale);
+    const f32 shadowOffset = std::round(1.0f * dpiScale);
 
     // console font does not scale with DPI
     static const i32 runtimeConfigHeight = (i32)(engine->getConsoleFont()->getHeight() * 1.25f);
@@ -1803,7 +1803,7 @@ void HUD::drawStatistics(const HUDStats &s) {
 
             g->translate(xOffset, yOffset);
 
-            g->drawString(font, text, TextShadow{.col_text = textColor, .col_shadow = shadowColor, .offs_px = 1});
+            g->drawString(font, text, TextShadow{.col_text = textColor, .col_shadow = shadowColor, .offs_px = 1.f});
 
             g->translate((-xOffset), (-yOffset) + yDelta);
         };
